Dockerfile: Only build QEMU for ppc
We require QEMU 3.0 for booting PowerPC kernels because of this commit.

https://github.com/qemu/qemu/commit/6b37554458a1e8cfe8f91ef0beac82e8de8f85bb

Joel brings up a good point that we should try to get this fixed in
upstream Debian. To minimize our delta, let's only build the updated
QEMU binary that we need and use Debian's official arm{32,64} and x86_64
packages.

Eventually, we will be able to kill off this entire section and only use
official packages (which would help a lot for creating more images for
different architectures).

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>