dockerimage: Move to Ubuntu Focal
Since moving to Debian unstable, there have been two independent errors
that have manifested during 'apt upgrade/install', resulting in no image
update. This is a little bit too unstable for my liking.  We need these
images to be up to date so we get accurate CI results.

https://travis-ci.com/github/ClangBuiltLinux/dockerimage/jobs/319302223
https://travis-ci.com/github/ClangBuiltLinux/dockerimage/jobs/319714044

We could try to move to Debian testing but it does not have official
support on apt.llvm.org so we risk dependency issues trying to use the
Debian unstable apt.llvm.org repos with a Debian testing base.

Ubuntu Focal gives us QEMU 4.2 (the reason to move to Debian unstable in
the first place) and it is a week away from its official release so we
should not be running into any instability with it. Additionally, it is
an LTS release so hopefully we won't need to move from it for a while.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>